requirements.txt
Flask==0.12.2
google-cloud-storage==1.6.0
gunicorn==19.7.1
cryptography==2.0